extract instructions
